+++ example/test/example_smoke_test.dart
@@ -1,6 +1,8 @@
-import 'package:auto_shimmer_animate_example/main.dart';
 import 'package:auto_shimmer_animate/src/core/utils/skeleton_box.dart';
 import 'package:flutter_test/flutter_test.dart';
+
+// ignore: avoid_relative_lib_imports
+import '../lib/main.dart';
 
 void main() {
   test('example app can be constructed', () {